Upgrade parent to 14.10 #54
* upgraded parent version to 14.10-1
* imported all documents from a 14.10 XWiki instance
* refactored configuration classes
* refactored collabora buttons
* tested functionalities
