+++ README.md
@@ -6,7 +6,7 @@ Integration of Collabora Online in XWiki. View, edit and collaborate real-time o
 * [Documentation](https://store.xwiki.com/xwiki/bin/view/Extension/Collabora%20Connector%20Application%20%28Pro%29/)
 * Communication: [Mailing List](http://dev.xwiki.org/xwiki/bin/view/Community/MailingLists), [IRC]( http://dev.xwiki.org/xwiki/bin/view/Community/IRC)
 * [Development Practices](http://dev.xwiki.org)
-* Minimal XWiki version supported: XWiki 13.10
+* Minimal XWiki version supported: XWiki 14.10
 * License: LGPL 2.1+
 * Translations: N/A
 * Sonar Dashboard: N/A
