+++ application-collabora-default/src/main/resources/META-INF/components.txt
@@ -4,6 +4,5 @@ com.xwiki.collabora.internal.AttachmentManager
 com.xwiki.collabora.internal.DiscoveryManager
 com.xwiki.collabora.internal.UserManager
 com.xwiki.collabora.internal.configuration.CollaboraConfigurationSource
-com.xwiki.collabora.internal.configuration.MainCollaboraConfigurationSource
 com.xwiki.collabora.internal.configuration.DefaultCollaboraConfiguration
 com.xwiki.collabora.script.CollaboraScriptService
