+++ application-collabora-ui/src/main/resources/Collabora/Code/Configuration.xml
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 -->
 
-<xwikidoc version="1.4" reference="Collabora.Code.Configuration" locale="">
+<xwikidoc version="1.5" reference="Collabora.Code.Configuration" locale="">
   <web>Collabora.Code</web>
   <name>Configuration</name>
   <language/>
