+++ application-collabora-ui/src/main/resources/Collabora/WebHome.xml
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
 -->
 
-<xwikidoc version="1.4" reference="Collabora.WebHome" locale="">
+<xwikidoc version="1.5" reference="Collabora.WebHome" locale="">
   <web>Collabora</web>
   <name>WebHome</name>
   <language/>
